chore: moved deep equal check to utils
